+++ src/game/player/PlayerTemplate.json
