+++ src/game/vue-game-plugin.js
@@ -2,11 +2,11 @@ import gsMan from "./state/gsMan";
 // import store from "../vuex/store";
 import gData from "./state/data";
 // import CommandParser from "./Commands/CommandParser";
-import GenerateName from "./Generators/NameGenerator";
+import GenerateName from "./utils/nameGenerator";
 import display from "./overview/Display";
 
 export default {
-  install: (Vue) => {
+  install: Vue => {
     Vue.prototype.$game = {
       start() {
         gsMan.StartGame();
@@ -20,12 +20,12 @@ export default {
         console.log("Parsing " + command);
       },
       worldExists() {
-        return (gData.world !== null);
+        return gData.world !== null;
       },
       displaySplash() {
         return gData.displaySplash;
       },
       display: display,
-    }
-  }
-}
+    };
+  },
+};
